class-based views

+++ store/views.py
@@ -1,45 +1,50 @@
 from django.shortcuts import get_object_or_404
 from django.http import HttpResponse
+from django.db.models import Count
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+from rest_framework.views import APIView
 from rest_framework import status
 from .models import Product, Collection
 from .serializers import ProductSerializer, CollectionSerializer
 
-# Create your views here.
-@api_view(['GET', 'POST'])
-def product_list(request):
-    if request.method == 'GET':
+class ProductList(APIView):
+    def get(self, request):
         queryset = Product.objects.select_related('collection').all()
         serializer = ProductSerializer(queryset, many=True, context={'request': request})
         return Response(serializer.data)
-    elif request.method == 'POST':
+    
+    def post(self, request):
         serializer = ProductSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
     
-@api_view(['GET', 'PUT', 'DELETE'])
-def product_detail(request, id):
-    product = get_object_or_404(Product, pk=id)
-    if request.method == 'GET':
+class ProductDetail(APIView):
+    def get(self, request, id):
+        product = get_object_or_404(Product, pk=id)
         serializer = ProductSerializer(product)
         return Response(serializer.data)
-    elif request.method == 'PUT':
+    
+    def put(self, request, id):
+        product = get_object_or_404(Product, pk=id)
         serializer = ProductSerializer(product, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
-    elif request.method == 'DELETE':
+
+    def delete(self, request, id):
+        product = get_object_or_404(Product, pk=id)
         if product.orderitems.count() > 0:
             return Response({'error': 'Product cannot be deleted because it is associated with an order item'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+    
 
 @api_view(['GET', 'POST'])
 def collection_list(request):
     if request.method == 'GET':
-        queryset = Collection.objects.all()
+        queryset = Collection.objects.annotate(products_count=Count('products')).all()
         serializer = CollectionSerializer(queryset, many=True)
         return Response(serializer.data)
     elif request.method == 'POST':
@@ -49,8 +54,8 @@ def collection_list(request):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
     
 @api_view(['GET', 'PUT', 'DELETE'])
-def collection_detail(request, id):
-    collection = get_object_or_404(Collection, pk=id)
+def collection_detail(request, pk):
+    collection = get_object_or_404(Collection.objects.annotate(products_count=Count('products')), pk=pk)
     if request.method == 'GET':
         serializer = CollectionSerializer(collection)
         return Response(serializer.data)
@@ -60,5 +65,7 @@ def collection_detail(request, id):
         serializer.save()
         return Response(serializer.data)
     elif request.method == 'DELETE':
+        if collection.products.count() > 0:
+            return Response({'error': 'Collection cannot be deleted because it includes one or more products.'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
